refactor: arrumei o nome do package de entites para entities
