--- src/main/java/com/br/aerotool/domain/entites/ToolItemStatus.java
@@ -1,4 +1,0 @@
-package com.br.aerotool.domain.entites;
-
-public enum ToolItemStatus {
-}
