+++ src/main/java/com/br/aerotool/domain/entities/Tool.java
@@ -1,4 +1,4 @@
-package com.br.aerotool.domain.entites;
+package com.br.aerotool.domain.entities;
 
 
 import java.util.Objects;
