+++ src/main/java/com/br/aerotool/domain/entities/ToolItemStatus.java
@@ -1,0 +1,4 @@
+package com.br.aerotool.domain.entities;
+
+public enum ToolItemStatus {
+}
